Fixed unit conversion bug
When weather was displayed in Fahrenheit and new city/current locataion was searched, the weather would be updated in Celsius but the unit would stay "Fahrenheit". This caused an unit conversion error because clicking on "Fahrenheit" would trigger the calculation into "Celsius" despite the numerical temperature displayed was already in Celsius.